# openapi.yaml
openapi: 3.0.3
info:
  title: UseZoracle Telegram Bot API
  description: API for managing cryptocurrency accounts, checking balances, and performing transactions using the Coinbase CDP SDK
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://usezoracle-telegrambot.up.railway.app
    description: Production server
...
paths:
...
  /api/swaps/tokens/{network}:
    get:
      summary: Get common token addresses for a network
      description: Returns a list of common token addresses for the specified network
      tags:
        - Swaps
      parameters:
        - name: network
          in: path
          required: true
          description: The network to get token addresses for
          schema:
            type: string
            enum: [base, ethereum, base-sepolia]
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/swaps/tokens/base"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/swaps/tokens/base')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenListResponse"
        "400":
          description: Invalid network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/swaps/price:
    get:
      operationId: getSwapPriceEstimate
      summary: Get swap price estimate
      description: Returns price estimate for swapping tokens. This is for display purposes only and doesn't reserve funds.
      tags:
        - Swaps
      parameters:
        - name: accountName
          in: query
          required: true
          description: Name of the account to use for the swap
          schema:
            type: string
        - name: fromToken
          in: query
          required: true
          description: Token address to swap from (or ETH for native Ethereum)
          schema:
            type: string
        - name: toToken
          in: query
          required: true
          description: Token address to swap to (or ETH for native Ethereum)
          schema:
            type: string
        - name: fromAmount
          in: query
          required: true
          description: Amount to swap in base units (e.g. wei for ETH)
          schema:
            type: string
        - name: network
          in: query
          required: false
          description: Network to perform the swap on
          schema:
            type: string
            enum: [base, ethereum]
            default: base
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/swaps/price?accountName=myAccount&fromToken=0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE&toToken=0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913&fromAmount=1000000000000000000&network=base"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/swaps/price?accountName=myAccount&fromToken=0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE&toToken=0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913&fromAmount=1000000000000000000&network=base')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapPriceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/swaps/execute:
    post:
      operationId: executeSwap
      summary: Execute token swap
      description: |
        Executes a swap between tokens with automatic token allowance management.

        **Important Notes:**
        - Automatically handles token allowance checking and approval for ERC20 tokens
        - Supports ERC20 to ERC20, ETH to ERC20, and ERC20 to ETH swaps
        - Works for all token types (ETH, USDC, WETH, etc.)
        - No fees applied - direct swap execution
      tags:
        - Swaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountName
                - fromToken
                - toToken
                - fromAmount
              properties:
                accountName:
                  type: string
                  description: Name of the account to execute the swap from
                  example: "myAccount"
                fromToken:
                  type: string
                  description: Token address to swap from (or ETH for native Ethereum)
                  example: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"
                toToken:
                  type: string
                  description: Token address to swap to (or ETH for native Ethereum)
                  example: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
                fromAmount:
                  type: string
                  description: Amount to swap in base units (e.g. wei for ETH)
                  example: "1000000000000000000"
                slippageBps:
                  type: number
                  description: Slippage tolerance in basis points (1 bps = 0.01%)
                  default: 100
                  minimum: 0
                  maximum: 5000
                  example: 100
                network:
                  type: string
                  description: Network to perform the swap on
                  enum: [base, ethereum]
                  default: base
                  example: "base"
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST "http://localhost:3000/api/swaps/execute" \
              -H "Content-Type: application/json" \
              -d '{"accountName":"myAccount","fromToken":"0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE","toToken":"0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913","fromAmount":"1000000000000000000","slippageBps":100,"network":"base"}'
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/swaps/execute', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({
                accountName: 'myAccount',
                fromToken: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
                toToken: '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913',
                fromAmount: '1000000000000000000',
                slippageBps: 100,
                network: 'base'
              })
            })
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwapExecutionResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/tokens/{contractAddress}:
    get:
      operationId: getBasicTokenInfo
      summary: Get token information
      description: |
        Retrieve basic token information (name, symbol, decimals) using the CDP service.

        For market data (price, supply, pools), use the Token Details endpoints:
        - `GET /api/token-details/tokens/{address}`
        - `GET /api/token-details/tokens/{address}/with-pools`
      tags:
        - Tokens
      parameters:
        - in: path
          name: contractAddress
          required: true
          schema:
            type: string
          description: Token contract address
          example: "0x4200000000000000000000000000000000000006"
        - in: query
          name: network
          schema:
            type: string
            enum: [base, base-sepolia, ethereum]
          description: Network to query (defaults to base)
          example: "base"
      x-code-samples:
        - lang: curl
          source: |
            curl -X GET "http://localhost:3000/api/tokens/0x4200000000000000000000000000000000000006?network=base"
        - lang: javascript
          source: |
            fetch('http://localhost:3000/api/tokens/0x4200000000000000000000000000000000000006?network=base')
              .then(response => response.json())
              .then(data => console.log(data));
      responses:
        "200":
          description: Token information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Human-readable message about the operation
        data:
          type: object
          description: Response data
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp when the error occurred
...
    TokenListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Token addresses for base network"
        data:
          type: object
          additionalProperties:
            type: string
          example:
            ETH: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE"
            WETH: "0x4200000000000000000000000000000000000006"
            USDC: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
            USDT: "0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA"
    SwapPriceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Swap price estimated successfully"
        data:
          type: object
          properties:
            liquidityAvailable:
              type: boolean
              description: Whether liquidity is available for this swap
              example: true
            fromAmount:
              type: string
              description: Amount of input token (in base units)
              example: "1000000000000000000"
            toAmount:
              type: string
              description: Amount of output token user will receive (in base units)
              example: "1800000000"
            minToAmount:
              type: string
              description: Minimum amount of output token considering slippage (in base units)
              example: "1782000000"
            expectedOutputFormatted:
              type: string
              description: Formatted amount user will receive (human-readable)
              example: "1800.0"
            minOutputFormatted:
              type: string
              description: Formatted minimum amount considering slippage (human-readable)
              example: "1782.0"
            exchangeRate:
              type: string
              description: Exchange rate between fromToken and toToken
              example: "1800.00000000"
    SwapExecutionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Swap executed successfully"
        data:
          type: object
          properties:
            transactionHash:
              type: string
              description: Hash of the executed swap transaction
              example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
            fromAmount:
              type: string
              description: Amount of input token (in base units)
              example: "1000000000000000000"
            network:
              type: string
              description: Network the swap was executed on
              example: "base"
            blockNumber:
              type: number
              description: Block number where the transaction was confirmed
              example: 33806419
            gasUsed:
              type: string
              description: Gas used for the transaction
              example: "1417788"
            status:
              type: string
              description: Transaction status
              example: "Success"
            transactionExplorer:
              type: string
              description: URL to view the transaction on the blockchain explorer
              example: "https://basescan.org/tx/0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
...